requirements.txt
asgiref==3.5.2
celery==4.4.1
dj-database-url==0.5.0
Django==3.2
django-environ==0.9.0
django-queryinspect==1.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.3
psycopg2-binary==2.9.5
pytz==2022.6
redis==3.4.1
slack-sdk==3.19.2
sqlparse==0.4.3
